move addon setup code to module
